feat(ui): add genre filter section to MovieList
This part is a small section of the MovieList component, which helps
user to navigate to find the genre they want to choose. User can use
mouse to scroll horizontally or use the navigation button on the right.

Besides, since I use position relative for the genre filter, I have to
change the index of the header so that it floats on top of everything
else including the genre filter.