refactor(Header): remove Logo component
Since I will probably not use Logo else where in the application, I
remove the Logo component and add the logo directly to the Header in
terms of element.

Besides, I also modify the accessibility attributes and content of
Header so that it can be as accessibility friendly as possible.