feat(route): add routing setup
Besides adding the routing setup, I also change the structure of App.jsx
a little bit. Now the outletContainer will not define the structure of
the page. Instead, components in /page will do this job. The
outletContainer in App.jsx only plays its role as a container of the
main section of the site.