fix(trailers): add minor fixes to trailer sections
Since the wanted behaviour is when clicking a trailer preview on the
navigation bar, the bar should not disappear, I should only add
stopPropagation to that ul element instead of the whole Watching Trailer
section.

Besides, I also modify the key prop of trailer item in Trailers.jsx.
The name of the prop was changed long time ago, but was missed in this
component.