fix(search): fix the unknown movie searching
When searching for a movie that does not exist, the page will display
an error. Previously, when user search for something else, the page will
not do anything, the user has to refresh to get results. But now they
do not have to do it anymore since I clear the error message before
fetching new data.